Implement CreateRoom controller and repository; rename room_id to roomId in Room entity

+++ bidverse/src/main/java/com/example/bidverse/Repository/RoomRepository.java
@@ -1,7 +1,7 @@
 package com.example.bidverse.Repository;
 import com.example.bidverse.Entity.Room;
 import org.springframework.data.repository.Repository;
-public interface RoomRepository extends Repository<Room,Integer> {
+public interface RoomRepository extends Repository<Room,Long> {
     Room findById(Long roomId);
     void save(Room room);
 }
